requirements.txt
google-generativeai
crewai
crewai-tools
python-dotenv
langchain-google-genai
langchain-community

